+++ pkgs/dev/src/vitest.config.ts
@@ -1,7 +1,11 @@
 import { defineConfig } from "vitest/config";
-export default defineConfig({
+
+export const vitestConfig = defineConfig({
   test: {
     isolate: false,
     passWithNoTests: true,
+    testTimeout: 100,
   },
 });
+
+export default vitestConfig;